OcAppleKernelLib: Compilation fixes

+++ Library/OcAppleKernelLib/PrelinkedKext.c
@@ -148,7 +148,7 @@ InternalCreatePrelinkedKext (
   }
 
   if (Prelinked != NULL
-    && !MachoInitializeContext (&NewKext->Context.MachContext, &Prelinked->Prelinked[SourceBase], SourceSize)) {
+    && !MachoInitializeContext (&NewKext->Context.MachContext, &Prelinked->Prelinked[SourceBase], (UINT32)SourceSize)) {
     FreePool (NewKext);
     return NULL;
   }
